Move token logic to separate transport wrapper
This patch moves the token management logic to the new authentication
transport wrapper.
