+++ h2c_test.go
@@ -25,10 +25,11 @@ import (
 	"path/filepath"
 	"time"
 
-	. "github.com/onsi/ginkgo" // nolint
-	. "github.com/onsi/gomega" // nolint
-
 	"github.com/onsi/gomega/ghttp"
+
+	. "github.com/onsi/ginkgo"                         // nolint
+	. "github.com/onsi/gomega"                         // nolint
+	. "github.com/openshift-online/ocm-sdk-go/testing" // nolint
 )
 
 var _ = Describe("H2C", func() {
@@ -40,8 +41,8 @@ var _ = Describe("H2C", func() {
 
 	BeforeEach(func() {
 		// Create the tokens:
-		accessToken = DefaultToken("Bearer", 5*time.Minute)
-		refreshToken = DefaultToken("Refresh", 10*time.Hour)
+		accessToken = MakeTokenString("Bearer", 5*time.Minute)
+		refreshToken = MakeTokenString("Refresh", 10*time.Hour)
 
 		// Create the OpenID server:
 		oidServer = MakeTCPServer()
