+++ unix_sockets_test.go
@@ -24,10 +24,11 @@ import (
 	"path/filepath"
 	"time"
 
-	. "github.com/onsi/ginkgo" // nolint
-	. "github.com/onsi/gomega" // nolint
-
 	"github.com/onsi/gomega/ghttp"
+
+	. "github.com/onsi/ginkgo"                         // nolint
+	. "github.com/onsi/gomega"                         // nolint
+	. "github.com/openshift-online/ocm-sdk-go/testing" // nolint
 )
 
 var _ = Describe("Unix sockets", func() {
@@ -39,8 +40,8 @@ var _ = Describe("Unix sockets", func() {
 
 	BeforeEach(func() {
 		// Create the tokens:
-		accessToken = DefaultToken("Bearer", 5*time.Minute)
-		refreshToken = DefaultToken("Refresh", 10*time.Hour)
+		accessToken = MakeTokenString("Bearer", 5*time.Minute)
+		refreshToken = MakeTokenString("Refresh", 10*time.Hour)
 
 		// Create the OpenID server:
 		oidServer, oidSocket = MakeUnixServer()
